+++ assets/javascripts/kitten/components/buttons/button-icon/index.js
@@ -1,37 +1,61 @@
-import React, { Component } from 'react'
+import React from 'react'
 import classNames from 'classnames'
+import { Button } from '../../../components/buttons/button/button'
+import styled from 'styled-components'
+import { pxToRem } from '../../../helpers/utils/typography'
 
-export class ButtonIcon extends Component {
-  render() {
-    const {
-      className,
-      tag,
-      modifier,
-      size,
-      withoutHover,
-      verticalArrow,
-      rounded,
-      ...others
-    } = this.props
+const StyledButton = styled(({ tag, ...props }) => (
+  <Button {...props} as={tag} />
+))`
+  svg,
+  .k-ButtonIconNext__svg {
+    display: block;
+    margin: 0;
+    padding: 0;
+    width: ${pxToRem(12)};
+    height: ${pxToRem(14)};
+    transition: fill 0.2s;
 
-    const buttonIconClassNames = classNames('k-ButtonIcon', className, {
-      [`k-ButtonIcon--${modifier}`]: modifier,
-      [`k-ButtonIcon--${size}`]: size,
-      'k-ButtonIcon--withoutHover': withoutHover,
-      'k-ButtonIcon--verticalArrow': verticalArrow,
-      'k-ButtonIcon--rounded': rounded,
-    })
+    pointer-events: none;
 
-    // Adds keyboard accessibility to `<a>`
-    const tabindex = tag == 'a' && !this.props.href ? 0 : null
-
-    const Tag = tag
-
-    return (
-      <Tag className={buttonIconClassNames} tabIndex={tabindex} {...others} />
-    )
+    .k-ButtonIconNext__svg__stroke {
+      transition: stroke 0.2s;
+    }
   }
-}
+
+  &.k-ButtonIconNext--nano svg,
+  &.k-ButtonIconNext--nano .k-ButtonIconNext__svg {
+    width: ${pxToRem(8)};
+    height: ${pxToRem(8)};
+  }
+
+  .k-ButtonIconNext--withoutHover {
+    pointer-events: none;
+  }
+`
+
+export const ButtonIcon = ({
+  className,
+  tag,
+  size,
+  withoutHover,
+  verticalArrow,
+  ...others
+}) => (
+  <StyledButton
+    className={classNames('k-ButtonIconNext', className, {
+      [`k-ButtonIconNext--${size}`]: size,
+      'k-ButtonIconNext--withoutHover': withoutHover,
+      'k-ButtonIconNext--verticalArrow': verticalArrow,
+    })}
+    nano={size === 'nano'}
+    micro={size === 'micro'}
+    tiny={size === 'tiny'}
+    big={size === 'big'}
+    {...others}
+    icon
+  />
+)
 
 ButtonIcon.defaultProps = {
   tag: 'button',